chore: DEV-2887: refactor TextArea to its own component
