+++ src/components/Comments/Comments.styl
@@ -1,4 +1,4 @@
 .comments
-    --comments-padding 12px 16px
+    --comments-padding 0 16px 12px  
 
     padding var(--comments-padding)
